separate game type
